refactor(seam): split SEAM binary, HIL, and opcode tables
