+++ crates/music_seam/src/opcode.rs
@@ -1,1150 +1,9 @@
-use crate::instruction::OperandShape;
+mod model;
+mod table;
 
-#[derive(Debug, Clone, Copy, PartialEq, Eq, Hash)]
-pub enum OpcodeFamily {
-    Core,
-    Storage,
-    Scalar,
-    Branch,
-    Call,
-    Object,
-    Type,
-    Ref,
-    Effect,
-    Module,
-}
-
-#[derive(Debug, Clone, Copy, PartialEq, Eq, Hash)]
-pub enum Opcode {
-    Nop,
-    Trap,
-    Pop,
-    Dup,
-    Swap,
-    LdUnit,
-    LdTrue,
-    LdFalse,
-    LdNull,
-    LdC,
-    LdCI4,
-    LdCI8,
-    LdCF4,
-    LdCF8,
-    LdStr,
-    LdArg,
-    StArg,
-    LdLoc,
-    StLoc,
-    LdGlob,
-    StGlob,
-    LdFld,
-    StFld,
-    LdFldA,
-    InitObj,
-    LdDflt,
-    Neg,
-    Add,
-    Sub,
-    Mul,
-    AddOvf,
-    SubOvf,
-    MulOvf,
-    DivS,
-    DivU,
-    RemS,
-    RemU,
-    And,
-    Or,
-    Xor,
-    Not,
-    Shl,
-    ShrS,
-    ShrU,
-    Rotl,
-    Rotr,
-    Clz,
-    Ctz,
-    Popcnt,
-    Ceq,
-    Cne,
-    CltS,
-    CltU,
-    CgtS,
-    CgtU,
-    CleS,
-    CleU,
-    CgeS,
-    CgeU,
-    Br,
-    BrTrue,
-    BrFalse,
-    BrTbl,
-    Leave,
-    Ret,
-    Throw,
-    Rethrow,
-    Unreach,
-    Call,
-    CallInd,
-    CallVirt,
-    CallIface,
-    CallDyn,
-    CallFfi,
-    TailCall,
-    TailCallInd,
-    TailCallVirt,
-    TailCallIface,
-    TailCallDyn,
-    TailCallFfi,
-    LdFn,
-    NewFn,
-    LdVirtFn,
-    LdIfaceFn,
-    LdDynFn,
-    LdFfi,
-    NewObj,
-    NewArr,
-    ArrNew,
-    LdElem,
-    StElem,
-    LdElemA,
-    LdLen,
-    Slice,
-    CpObj,
-    DupObj,
-    DropObj,
-    LdType,
-    TypeOf,
-    IsInst,
-    Cast,
-    Conv,
-    ConvOvf,
-    ConvS,
-    ConvU,
-    ConvOvfS,
-    ConvOvfU,
-    Box,
-    Unbox,
-    SizeOf,
-    AlignOf,
-    LdTok,
-    LdInd,
-    StInd,
-    RefAny,
-    EndRef,
-    Pin,
-    Unpin,
-    LdAddr,
-    PtrEq,
-    PtrCast,
-    HdlPush,
-    HdlPop,
-    Raise,
-    Resume,
-    DropCont,
-    LdImp,
-    LdExp,
-    MdlLoad,
-    MdlGet,
-}
-
-#[derive(Debug, Clone, Copy, PartialEq, Eq, Hash)]
-pub enum OpcodeWire {
-    Core(u8),
-    Extended(u16),
-}
-
-const EXTENDED_OPCODE_PREFIX: u8 = 0xFF;
-
-impl Opcode {
-    #[must_use]
-    pub fn family(self) -> OpcodeFamily {
-        self.info().family
-    }
-    #[must_use]
-    pub fn mnemonic(self) -> &'static str {
-        self.info().mnemonic
-    }
-    #[must_use]
-    pub fn operand_shape(self) -> OperandShape {
-        self.info().operand_shape
-    }
-    #[must_use]
-    pub fn wire_code(self) -> u16 {
-        match self.info().wire {
-            OpcodeWire::Core(code) => u16::from(code),
-            OpcodeWire::Extended(code) => code,
-        }
-    }
-    #[must_use]
-    pub fn wire(self) -> OpcodeWire {
-        self.info().wire
-    }
-    #[must_use]
-    pub const fn extended_opcode_prefix() -> u8 {
-        EXTENDED_OPCODE_PREFIX
-    }
-    #[must_use]
-    pub fn from_mnemonic(text: &str) -> Option<Self> {
-        OPCODE_INFOS
-            .iter()
-            .find(|info| info.mnemonic == text)
-            .map(|info| info.opcode)
-    }
-    #[must_use]
-    pub fn from_wire_code(code: u16) -> Option<Self> {
-        OPCODE_INFOS
-            .iter()
-            .find(|info| match info.wire {
-                OpcodeWire::Core(core_wire) => u16::from(core_wire) == code,
-                OpcodeWire::Extended(extended) => extended == code,
-            })
-            .map(|info| info.opcode)
-    }
-    fn info(self) -> &'static OpcodeInfo {
-        OPCODE_INFOS
-            .iter()
-            .find(|info| info.opcode == self)
-            .unwrap_or(&OPCODE_INFOS[0])
-    }
-}
-
-struct OpcodeInfo {
-    opcode: Opcode,
-    family: OpcodeFamily,
-    mnemonic: &'static str,
-    operand_shape: OperandShape,
-    wire: OpcodeWire,
-}
-
-const fn opcode_info(
-    opcode: Opcode,
-    family: OpcodeFamily,
-    mnemonic: &'static str,
-    operand_shape: OperandShape,
-    code: u8,
-) -> OpcodeInfo {
-    OpcodeInfo {
-        opcode,
-        family,
-        mnemonic,
-        operand_shape,
-        wire: OpcodeWire::Core(code),
-    }
-}
-
-const OPCODE_INFOS: &[OpcodeInfo] = &[
-    opcode_info(
-        Opcode::Nop,
-        OpcodeFamily::Core,
-        "nop",
-        OperandShape::None,
-        0x00,
-    ),
-    opcode_info(
-        Opcode::Trap,
-        OpcodeFamily::Core,
-        "trap",
-        OperandShape::String,
-        0x01,
-    ),
-    opcode_info(
-        Opcode::Pop,
-        OpcodeFamily::Core,
-        "pop",
-        OperandShape::None,
-        0x02,
-    ),
-    opcode_info(
-        Opcode::Dup,
-        OpcodeFamily::Core,
-        "dup",
-        OperandShape::None,
-        0x03,
-    ),
-    opcode_info(
-        Opcode::Swap,
-        OpcodeFamily::Core,
-        "swap",
-        OperandShape::None,
-        0x04,
-    ),
-    opcode_info(
-        Opcode::LdUnit,
-        OpcodeFamily::Core,
-        "ld.unit",
-        OperandShape::None,
-        0x05,
-    ),
-    opcode_info(
-        Opcode::LdTrue,
-        OpcodeFamily::Core,
-        "ld.true",
-        OperandShape::None,
-        0x06,
-    ),
-    opcode_info(
-        Opcode::LdFalse,
-        OpcodeFamily::Core,
-        "ld.false",
-        OperandShape::None,
-        0x07,
-    ),
-    opcode_info(
-        Opcode::LdNull,
-        OpcodeFamily::Core,
-        "ld.null",
-        OperandShape::Type,
-        0x08,
-    ),
-    opcode_info(
-        Opcode::LdC,
-        OpcodeFamily::Core,
-        "ld.c",
-        OperandShape::Constant,
-        0x09,
-    ),
-    opcode_info(
-        Opcode::LdCI4,
-        OpcodeFamily::Core,
-        "ld.c.i4",
-        OperandShape::I16,
-        0x0A,
-    ),
-    opcode_info(
-        Opcode::LdCI8,
-        OpcodeFamily::Core,
-        "ld.c.i8",
-        OperandShape::I16,
-        0x0B,
-    ),
-    opcode_info(
-        Opcode::LdCF4,
-        OpcodeFamily::Core,
-        "ld.c.f4",
-        OperandShape::I16,
-        0x0C,
-    ),
-    opcode_info(
-        Opcode::LdCF8,
-        OpcodeFamily::Core,
-        "ld.c.f8",
-        OperandShape::I16,
-        0x0D,
-    ),
-    opcode_info(
-        Opcode::LdStr,
-        OpcodeFamily::Core,
-        "ld.str",
-        OperandShape::String,
-        0x0E,
-    ),
-    opcode_info(
-        Opcode::LdArg,
-        OpcodeFamily::Storage,
-        "ld.arg",
-        OperandShape::Local,
-        0x10,
-    ),
-    opcode_info(
-        Opcode::StArg,
-        OpcodeFamily::Storage,
-        "st.arg",
-        OperandShape::Local,
-        0x11,
-    ),
-    opcode_info(
-        Opcode::LdLoc,
-        OpcodeFamily::Storage,
-        "ld.loc",
-        OperandShape::Local,
-        0x12,
-    ),
-    opcode_info(
-        Opcode::StLoc,
-        OpcodeFamily::Storage,
-        "st.loc",
-        OperandShape::Local,
-        0x13,
-    ),
-    opcode_info(
-        Opcode::LdGlob,
-        OpcodeFamily::Storage,
-        "ld.glob",
-        OperandShape::Global,
-        0x14,
-    ),
-    opcode_info(
-        Opcode::StGlob,
-        OpcodeFamily::Storage,
-        "st.glob",
-        OperandShape::Global,
-        0x15,
-    ),
-    opcode_info(
-        Opcode::LdFld,
-        OpcodeFamily::Storage,
-        "ld.fld",
-        OperandShape::I16,
-        0x16,
-    ),
-    opcode_info(
-        Opcode::StFld,
-        OpcodeFamily::Storage,
-        "st.fld",
-        OperandShape::I16,
-        0x17,
-    ),
-    opcode_info(
-        Opcode::LdFldA,
-        OpcodeFamily::Storage,
-        "ld.fld.a",
-        OperandShape::I16,
-        0x18,
-    ),
-    opcode_info(
-        Opcode::InitObj,
-        OpcodeFamily::Storage,
-        "init.obj",
-        OperandShape::Type,
-        0x19,
-    ),
-    opcode_info(
-        Opcode::LdDflt,
-        OpcodeFamily::Storage,
-        "ld.dflt",
-        OperandShape::Type,
-        0x1A,
-    ),
-    opcode_info(
-        Opcode::Neg,
-        OpcodeFamily::Scalar,
-        "neg",
-        OperandShape::None,
-        0x20,
-    ),
-    opcode_info(
-        Opcode::Add,
-        OpcodeFamily::Scalar,
-        "add",
-        OperandShape::None,
-        0x21,
-    ),
-    opcode_info(
-        Opcode::Sub,
-        OpcodeFamily::Scalar,
-        "sub",
-        OperandShape::None,
-        0x22,
-    ),
-    opcode_info(
-        Opcode::Mul,
-        OpcodeFamily::Scalar,
-        "mul",
-        OperandShape::None,
-        0x23,
-    ),
-    opcode_info(
-        Opcode::AddOvf,
-        OpcodeFamily::Scalar,
-        "add.ovf",
-        OperandShape::None,
-        0x24,
-    ),
-    opcode_info(
-        Opcode::SubOvf,
-        OpcodeFamily::Scalar,
-        "sub.ovf",
-        OperandShape::None,
-        0x25,
-    ),
-    opcode_info(
-        Opcode::MulOvf,
-        OpcodeFamily::Scalar,
-        "mul.ovf",
-        OperandShape::None,
-        0x26,
-    ),
-    opcode_info(
-        Opcode::DivS,
-        OpcodeFamily::Scalar,
-        "div.s",
-        OperandShape::None,
-        0x27,
-    ),
-    opcode_info(
-        Opcode::DivU,
-        OpcodeFamily::Scalar,
-        "div.u",
-        OperandShape::None,
-        0x28,
-    ),
-    opcode_info(
-        Opcode::RemS,
-        OpcodeFamily::Scalar,
-        "rem.s",
-        OperandShape::None,
-        0x29,
-    ),
-    opcode_info(
-        Opcode::RemU,
-        OpcodeFamily::Scalar,
-        "rem.u",
-        OperandShape::None,
-        0x2A,
-    ),
-    opcode_info(
-        Opcode::And,
-        OpcodeFamily::Scalar,
-        "and",
-        OperandShape::None,
-        0x2B,
-    ),
-    opcode_info(
-        Opcode::Or,
-        OpcodeFamily::Scalar,
-        "or",
-        OperandShape::None,
-        0x2C,
-    ),
-    opcode_info(
-        Opcode::Xor,
-        OpcodeFamily::Scalar,
-        "xor",
-        OperandShape::None,
-        0x2D,
-    ),
-    opcode_info(
-        Opcode::Not,
-        OpcodeFamily::Scalar,
-        "not",
-        OperandShape::None,
-        0x2E,
-    ),
-    opcode_info(
-        Opcode::Shl,
-        OpcodeFamily::Scalar,
-        "shl",
-        OperandShape::None,
-        0x2F,
-    ),
-    opcode_info(
-        Opcode::ShrS,
-        OpcodeFamily::Scalar,
-        "shr.s",
-        OperandShape::None,
-        0x30,
-    ),
-    opcode_info(
-        Opcode::ShrU,
-        OpcodeFamily::Scalar,
-        "shr.u",
-        OperandShape::None,
-        0x31,
-    ),
-    opcode_info(
-        Opcode::Rotl,
-        OpcodeFamily::Scalar,
-        "rotl",
-        OperandShape::None,
-        0x32,
-    ),
-    opcode_info(
-        Opcode::Rotr,
-        OpcodeFamily::Scalar,
-        "rotr",
-        OperandShape::None,
-        0x33,
-    ),
-    opcode_info(
-        Opcode::Clz,
-        OpcodeFamily::Scalar,
-        "clz",
-        OperandShape::None,
-        0x34,
-    ),
-    opcode_info(
-        Opcode::Ctz,
-        OpcodeFamily::Scalar,
-        "ctz",
-        OperandShape::None,
-        0x35,
-    ),
-    opcode_info(
-        Opcode::Popcnt,
-        OpcodeFamily::Scalar,
-        "popcnt",
-        OperandShape::None,
-        0x36,
-    ),
-    opcode_info(
-        Opcode::Ceq,
-        OpcodeFamily::Scalar,
-        "ceq",
-        OperandShape::None,
-        0x37,
-    ),
-    opcode_info(
-        Opcode::Cne,
-        OpcodeFamily::Scalar,
-        "cne",
-        OperandShape::None,
-        0x38,
-    ),
-    opcode_info(
-        Opcode::CltS,
-        OpcodeFamily::Scalar,
-        "clt.s",
-        OperandShape::None,
-        0x39,
-    ),
-    opcode_info(
-        Opcode::CltU,
-        OpcodeFamily::Scalar,
-        "clt.u",
-        OperandShape::None,
-        0x3A,
-    ),
-    opcode_info(
-        Opcode::CgtS,
-        OpcodeFamily::Scalar,
-        "cgt.s",
-        OperandShape::None,
-        0x3B,
-    ),
-    opcode_info(
-        Opcode::CgtU,
-        OpcodeFamily::Scalar,
-        "cgt.u",
-        OperandShape::None,
-        0x3C,
-    ),
-    opcode_info(
-        Opcode::CleS,
-        OpcodeFamily::Scalar,
-        "cle.s",
-        OperandShape::None,
-        0x3D,
-    ),
-    opcode_info(
-        Opcode::CleU,
-        OpcodeFamily::Scalar,
-        "cle.u",
-        OperandShape::None,
-        0x3E,
-    ),
-    opcode_info(
-        Opcode::CgeS,
-        OpcodeFamily::Scalar,
-        "cge.s",
-        OperandShape::None,
-        0x40,
-    ),
-    opcode_info(
-        Opcode::CgeU,
-        OpcodeFamily::Scalar,
-        "cge.u",
-        OperandShape::None,
-        0x41,
-    ),
-    opcode_info(
-        Opcode::Br,
-        OpcodeFamily::Branch,
-        "br",
-        OperandShape::Label,
-        0x42,
-    ),
-    opcode_info(
-        Opcode::BrTrue,
-        OpcodeFamily::Branch,
-        "br.true",
-        OperandShape::Label,
-        0x43,
-    ),
-    opcode_info(
-        Opcode::BrFalse,
-        OpcodeFamily::Branch,
-        "br.false",
-        OperandShape::Label,
-        0x44,
-    ),
-    opcode_info(
-        Opcode::BrTbl,
-        OpcodeFamily::Branch,
-        "br.tbl",
-        OperandShape::BranchTable,
-        0x45,
-    ),
-    opcode_info(
-        Opcode::Leave,
-        OpcodeFamily::Branch,
-        "leave",
-        OperandShape::Label,
-        0x46,
-    ),
-    opcode_info(
-        Opcode::Ret,
-        OpcodeFamily::Branch,
-        "ret",
-        OperandShape::None,
-        0x47,
-    ),
-    opcode_info(
-        Opcode::Throw,
-        OpcodeFamily::Branch,
-        "throw",
-        OperandShape::None,
-        0x48,
-    ),
-    opcode_info(
-        Opcode::Rethrow,
-        OpcodeFamily::Branch,
-        "rethrow",
-        OperandShape::None,
-        0x49,
-    ),
-    opcode_info(
-        Opcode::Unreach,
-        OpcodeFamily::Branch,
-        "unreach",
-        OperandShape::None,
-        0x4A,
-    ),
-    opcode_info(
-        Opcode::Call,
-        OpcodeFamily::Call,
-        "call",
-        OperandShape::Procedure,
-        0x50,
-    ),
-    opcode_info(
-        Opcode::CallInd,
-        OpcodeFamily::Call,
-        "call.ind",
-        OperandShape::None,
-        0x51,
-    ),
-    opcode_info(
-        Opcode::CallVirt,
-        OpcodeFamily::Call,
-        "call.virt",
-        OperandShape::Procedure,
-        0x52,
-    ),
-    opcode_info(
-        Opcode::CallIface,
-        OpcodeFamily::Call,
-        "call.iface",
-        OperandShape::Procedure,
-        0x53,
-    ),
-    opcode_info(
-        Opcode::CallDyn,
-        OpcodeFamily::Call,
-        "call.dyn",
-        OperandShape::Procedure,
-        0x54,
-    ),
-    opcode_info(
-        Opcode::CallFfi,
-        OpcodeFamily::Call,
-        "call.ffi",
-        OperandShape::Foreign,
-        0x55,
-    ),
-    opcode_info(
-        Opcode::TailCall,
-        OpcodeFamily::Call,
-        "tail.call",
-        OperandShape::Procedure,
-        0x56,
-    ),
-    opcode_info(
-        Opcode::TailCallInd,
-        OpcodeFamily::Call,
-        "tail.call.ind",
-        OperandShape::Procedure,
-        0x57,
-    ),
-    opcode_info(
-        Opcode::TailCallVirt,
-        OpcodeFamily::Call,
-        "tail.call.virt",
-        OperandShape::Procedure,
-        0x58,
-    ),
-    opcode_info(
-        Opcode::TailCallIface,
-        OpcodeFamily::Call,
-        "tail.call.iface",
-        OperandShape::Procedure,
-        0x59,
-    ),
-    opcode_info(
-        Opcode::TailCallDyn,
-        OpcodeFamily::Call,
-        "tail.call.dyn",
-        OperandShape::Procedure,
-        0x5A,
-    ),
-    opcode_info(
-        Opcode::TailCallFfi,
-        OpcodeFamily::Call,
-        "tail.call.ffi",
-        OperandShape::Foreign,
-        0x5B,
-    ),
-    opcode_info(
-        Opcode::LdFn,
-        OpcodeFamily::Call,
-        "ld.fn",
-        OperandShape::Procedure,
-        0x5C,
-    ),
-    opcode_info(
-        Opcode::NewFn,
-        OpcodeFamily::Call,
-        "new.fn",
-        OperandShape::WideProcedureCaptures,
-        0x5D,
-    ),
-    opcode_info(
-        Opcode::LdVirtFn,
-        OpcodeFamily::Call,
-        "ld.virt.fn",
-        OperandShape::Procedure,
-        0x5E,
-    ),
-    opcode_info(
-        Opcode::LdIfaceFn,
-        OpcodeFamily::Call,
-        "ld.iface.fn",
-        OperandShape::Procedure,
-        0x5F,
-    ),
-    opcode_info(
-        Opcode::LdDynFn,
-        OpcodeFamily::Call,
-        "ld.dyn.fn",
-        OperandShape::Procedure,
-        0x60,
-    ),
-    opcode_info(
-        Opcode::LdFfi,
-        OpcodeFamily::Call,
-        "ld.ffi",
-        OperandShape::Foreign,
-        0x61,
-    ),
-    opcode_info(
-        Opcode::NewObj,
-        OpcodeFamily::Object,
-        "new.obj",
-        OperandShape::TypeLen,
-        0x70,
-    ),
-    opcode_info(
-        Opcode::NewArr,
-        OpcodeFamily::Object,
-        "new.arr",
-        OperandShape::TypeLen,
-        0x71,
-    ),
-    opcode_info(
-        Opcode::ArrNew,
-        OpcodeFamily::Object,
-        "arr.new",
-        OperandShape::Type,
-        0x72,
-    ),
-    opcode_info(
-        Opcode::LdElem,
-        OpcodeFamily::Object,
-        "ld.elem",
-        OperandShape::None,
-        0x73,
-    ),
-    opcode_info(
-        Opcode::StElem,
-        OpcodeFamily::Object,
-        "st.elem",
-        OperandShape::None,
-        0x74,
-    ),
-    opcode_info(
-        Opcode::LdElemA,
-        OpcodeFamily::Object,
-        "ld.elem.a",
-        OperandShape::Type,
-        0x75,
-    ),
-    opcode_info(
-        Opcode::LdLen,
-        OpcodeFamily::Object,
-        "ld.len",
-        OperandShape::None,
-        0x76,
-    ),
-    opcode_info(
-        Opcode::Slice,
-        OpcodeFamily::Object,
-        "slice",
-        OperandShape::None,
-        0x77,
-    ),
-    opcode_info(
-        Opcode::CpObj,
-        OpcodeFamily::Object,
-        "cp.obj",
-        OperandShape::Type,
-        0x78,
-    ),
-    opcode_info(
-        Opcode::DupObj,
-        OpcodeFamily::Object,
-        "dup.obj",
-        OperandShape::Type,
-        0x79,
-    ),
-    opcode_info(
-        Opcode::DropObj,
-        OpcodeFamily::Object,
-        "drop.obj",
-        OperandShape::Type,
-        0x7A,
-    ),
-    opcode_info(
-        Opcode::LdType,
-        OpcodeFamily::Type,
-        "ld.type",
-        OperandShape::Type,
-        0x80,
-    ),
-    opcode_info(
-        Opcode::TypeOf,
-        OpcodeFamily::Type,
-        "type.of",
-        OperandShape::None,
-        0x81,
-    ),
-    opcode_info(
-        Opcode::IsInst,
-        OpcodeFamily::Type,
-        "is.inst",
-        OperandShape::Type,
-        0x82,
-    ),
-    opcode_info(
-        Opcode::Cast,
-        OpcodeFamily::Type,
-        "cast",
-        OperandShape::Type,
-        0x83,
-    ),
-    opcode_info(
-        Opcode::Conv,
-        OpcodeFamily::Type,
-        "conv",
-        OperandShape::Type,
-        0x84,
-    ),
-    opcode_info(
-        Opcode::ConvOvf,
-        OpcodeFamily::Type,
-        "conv.ovf",
-        OperandShape::Type,
-        0x85,
-    ),
-    opcode_info(
-        Opcode::ConvS,
-        OpcodeFamily::Type,
-        "conv.s",
-        OperandShape::Type,
-        0x86,
-    ),
-    opcode_info(
-        Opcode::ConvU,
-        OpcodeFamily::Type,
-        "conv.u",
-        OperandShape::Type,
-        0x87,
-    ),
-    opcode_info(
-        Opcode::ConvOvfS,
-        OpcodeFamily::Type,
-        "conv.ovf.s",
-        OperandShape::Type,
-        0x88,
-    ),
-    opcode_info(
-        Opcode::ConvOvfU,
-        OpcodeFamily::Type,
-        "conv.ovf.u",
-        OperandShape::Type,
-        0x89,
-    ),
-    opcode_info(
-        Opcode::Box,
-        OpcodeFamily::Type,
-        "box",
-        OperandShape::Type,
-        0x8A,
-    ),
-    opcode_info(
-        Opcode::Unbox,
-        OpcodeFamily::Type,
-        "unbox",
-        OperandShape::Type,
-        0x8B,
-    ),
-    opcode_info(
-        Opcode::SizeOf,
-        OpcodeFamily::Type,
-        "size.of",
-        OperandShape::Type,
-        0x8C,
-    ),
-    opcode_info(
-        Opcode::AlignOf,
-        OpcodeFamily::Type,
-        "align.of",
-        OperandShape::Type,
-        0x8D,
-    ),
-    opcode_info(
-        Opcode::LdTok,
-        OpcodeFamily::Type,
-        "ld.tok",
-        OperandShape::Type,
-        0x8E,
-    ),
-    opcode_info(
-        Opcode::LdInd,
-        OpcodeFamily::Ref,
-        "ld.ind",
-        OperandShape::Type,
-        0x90,
-    ),
-    opcode_info(
-        Opcode::StInd,
-        OpcodeFamily::Ref,
-        "st.ind",
-        OperandShape::Type,
-        0x91,
-    ),
-    opcode_info(
-        Opcode::RefAny,
-        OpcodeFamily::Ref,
-        "ref.any",
-        OperandShape::None,
-        0x92,
-    ),
-    opcode_info(
-        Opcode::EndRef,
-        OpcodeFamily::Ref,
-        "end.ref",
-        OperandShape::None,
-        0x93,
-    ),
-    opcode_info(
-        Opcode::Pin,
-        OpcodeFamily::Ref,
-        "pin",
-        OperandShape::None,
-        0x94,
-    ),
-    opcode_info(
-        Opcode::Unpin,
-        OpcodeFamily::Ref,
-        "unpin",
-        OperandShape::None,
-        0x95,
-    ),
-    opcode_info(
-        Opcode::LdAddr,
-        OpcodeFamily::Ref,
-        "ld.addr",
-        OperandShape::Type,
-        0x96,
-    ),
-    opcode_info(
-        Opcode::PtrEq,
-        OpcodeFamily::Ref,
-        "ptr.eq",
-        OperandShape::None,
-        0x97,
-    ),
-    opcode_info(
-        Opcode::PtrCast,
-        OpcodeFamily::Ref,
-        "ptr.cast",
-        OperandShape::Type,
-        0x98,
-    ),
-    opcode_info(
-        Opcode::HdlPush,
-        OpcodeFamily::Effect,
-        "hdl.push",
-        OperandShape::EffectId,
-        0xA0,
-    ),
-    opcode_info(
-        Opcode::HdlPop,
-        OpcodeFamily::Effect,
-        "hdl.pop",
-        OperandShape::None,
-        0xA1,
-    ),
-    opcode_info(
-        Opcode::Raise,
-        OpcodeFamily::Effect,
-        "raise",
-        OperandShape::Effect,
-        0xA2,
-    ),
-    opcode_info(
-        Opcode::Resume,
-        OpcodeFamily::Effect,
-        "resume",
-        OperandShape::None,
-        0xA3,
-    ),
-    opcode_info(
-        Opcode::DropCont,
-        OpcodeFamily::Effect,
-        "drop.cont",
-        OperandShape::None,
-        0xA4,
-    ),
-    opcode_info(
-        Opcode::LdImp,
-        OpcodeFamily::Module,
-        "ld.imp",
-        OperandShape::String,
-        0xB0,
-    ),
-    opcode_info(
-        Opcode::LdExp,
-        OpcodeFamily::Module,
-        "ld.exp",
-        OperandShape::String,
-        0xB1,
-    ),
-    opcode_info(
-        Opcode::MdlLoad,
-        OpcodeFamily::Module,
-        "mdl.load",
-        OperandShape::None,
-        0xB2,
-    ),
-    opcode_info(
-        Opcode::MdlGet,
-        OpcodeFamily::Module,
-        "mdl.get",
-        OperandShape::String,
-        0xB3,
-    ),
-];
+pub use model::{Opcode, OpcodeFamily, OpcodeWire};
+#[cfg(test)]
+use table::{opcode_info_count, opcode_infos};
 
 #[cfg(test)]
 mod tests;
