+++ crates/music_seam/src/opcode/model.rs
@@ -1,0 +1,153 @@
+#[derive(Debug, Clone, Copy, PartialEq, Eq, Hash)]
+pub enum OpcodeFamily {
+    Core,
+    Storage,
+    Scalar,
+    Branch,
+    Call,
+    Object,
+    Type,
+    Ref,
+    Effect,
+    Module,
+}
+
+#[derive(Debug, Clone, Copy, PartialEq, Eq, Hash)]
+pub enum Opcode {
+    Nop,
+    Trap,
+    Pop,
+    Dup,
+    Swap,
+    LdUnit,
+    LdTrue,
+    LdFalse,
+    LdNull,
+    LdC,
+    LdCI4,
+    LdCI8,
+    LdCF4,
+    LdCF8,
+    LdStr,
+    LdArg,
+    StArg,
+    LdLoc,
+    StLoc,
+    LdGlob,
+    StGlob,
+    LdFld,
+    StFld,
+    LdFldA,
+    InitObj,
+    LdDflt,
+    Neg,
+    Add,
+    Sub,
+    Mul,
+    AddOvf,
+    SubOvf,
+    MulOvf,
+    DivS,
+    DivU,
+    RemS,
+    RemU,
+    And,
+    Or,
+    Xor,
+    Not,
+    Shl,
+    ShrS,
+    ShrU,
+    Rotl,
+    Rotr,
+    Clz,
+    Ctz,
+    Popcnt,
+    Ceq,
+    Cne,
+    CltS,
+    CltU,
+    CgtS,
+    CgtU,
+    CleS,
+    CleU,
+    CgeS,
+    CgeU,
+    Br,
+    BrTrue,
+    BrFalse,
+    BrTbl,
+    Leave,
+    Ret,
+    Throw,
+    Rethrow,
+    Unreach,
+    Call,
+    CallInd,
+    CallVirt,
+    CallIface,
+    CallDyn,
+    CallFfi,
+    TailCall,
+    TailCallInd,
+    TailCallVirt,
+    TailCallIface,
+    TailCallDyn,
+    TailCallFfi,
+    LdFn,
+    NewFn,
+    LdVirtFn,
+    LdIfaceFn,
+    LdDynFn,
+    LdFfi,
+    NewObj,
+    NewArr,
+    ArrNew,
+    LdElem,
+    StElem,
+    LdElemA,
+    LdLen,
+    Slice,
+    CpObj,
+    DupObj,
+    DropObj,
+    LdType,
+    TypeOf,
+    IsInst,
+    Cast,
+    Conv,
+    ConvOvf,
+    ConvS,
+    ConvU,
+    ConvOvfS,
+    ConvOvfU,
+    Box,
+    Unbox,
+    SizeOf,
+    AlignOf,
+    LdTok,
+    LdInd,
+    StInd,
+    RefAny,
+    EndRef,
+    Pin,
+    Unpin,
+    LdAddr,
+    PtrEq,
+    PtrCast,
+    HdlPush,
+    HdlPop,
+    Raise,
+    Resume,
+    DropCont,
+    LdImp,
+    LdExp,
+    MdlLoad,
+    MdlGet,
+}
+
+#[derive(Debug, Clone, Copy, PartialEq, Eq, Hash)]
+pub enum OpcodeWire {
+    Core(u8),
+    Extended(u16),
+}
