+++ crates/music_seam/src/opcode/table/call.rs
@@ -1,0 +1,133 @@
+use crate::instruction::OperandShape;
+
+use super::super::{Opcode, OpcodeFamily};
+use super::info::{OpcodeInfo, opcode_info};
+
+pub(super) const OPCODES: &[OpcodeInfo] = &[
+    opcode_info(
+        Opcode::Call,
+        OpcodeFamily::Call,
+        "call",
+        OperandShape::Procedure,
+        0x50,
+    ),
+    opcode_info(
+        Opcode::CallInd,
+        OpcodeFamily::Call,
+        "call.ind",
+        OperandShape::None,
+        0x51,
+    ),
+    opcode_info(
+        Opcode::CallVirt,
+        OpcodeFamily::Call,
+        "call.virt",
+        OperandShape::Procedure,
+        0x52,
+    ),
+    opcode_info(
+        Opcode::CallIface,
+        OpcodeFamily::Call,
+        "call.iface",
+        OperandShape::Procedure,
+        0x53,
+    ),
+    opcode_info(
+        Opcode::CallDyn,
+        OpcodeFamily::Call,
+        "call.dyn",
+        OperandShape::Procedure,
+        0x54,
+    ),
+    opcode_info(
+        Opcode::CallFfi,
+        OpcodeFamily::Call,
+        "call.ffi",
+        OperandShape::Foreign,
+        0x55,
+    ),
+    opcode_info(
+        Opcode::TailCall,
+        OpcodeFamily::Call,
+        "tail.call",
+        OperandShape::Procedure,
+        0x56,
+    ),
+    opcode_info(
+        Opcode::TailCallInd,
+        OpcodeFamily::Call,
+        "tail.call.ind",
+        OperandShape::Procedure,
+        0x57,
+    ),
+    opcode_info(
+        Opcode::TailCallVirt,
+        OpcodeFamily::Call,
+        "tail.call.virt",
+        OperandShape::Procedure,
+        0x58,
+    ),
+    opcode_info(
+        Opcode::TailCallIface,
+        OpcodeFamily::Call,
+        "tail.call.iface",
+        OperandShape::Procedure,
+        0x59,
+    ),
+    opcode_info(
+        Opcode::TailCallDyn,
+        OpcodeFamily::Call,
+        "tail.call.dyn",
+        OperandShape::Procedure,
+        0x5A,
+    ),
+    opcode_info(
+        Opcode::TailCallFfi,
+        OpcodeFamily::Call,
+        "tail.call.ffi",
+        OperandShape::Foreign,
+        0x5B,
+    ),
+    opcode_info(
+        Opcode::LdFn,
+        OpcodeFamily::Call,
+        "ld.fn",
+        OperandShape::Procedure,
+        0x5C,
+    ),
+    opcode_info(
+        Opcode::NewFn,
+        OpcodeFamily::Call,
+        "new.fn",
+        OperandShape::WideProcedureCaptures,
+        0x5D,
+    ),
+    opcode_info(
+        Opcode::LdVirtFn,
+        OpcodeFamily::Call,
+        "ld.virt.fn",
+        OperandShape::Procedure,
+        0x5E,
+    ),
+    opcode_info(
+        Opcode::LdIfaceFn,
+        OpcodeFamily::Call,
+        "ld.iface.fn",
+        OperandShape::Procedure,
+        0x5F,
+    ),
+    opcode_info(
+        Opcode::LdDynFn,
+        OpcodeFamily::Call,
+        "ld.dyn.fn",
+        OperandShape::Procedure,
+        0x60,
+    ),
+    opcode_info(
+        Opcode::LdFfi,
+        OpcodeFamily::Call,
+        "ld.ffi",
+        OperandShape::Foreign,
+        0x61,
+    ),
+];
