+++ crates/music_seam/src/opcode/table/core.rs
@@ -1,0 +1,112 @@
+use crate::instruction::OperandShape;
+
+use super::super::{Opcode, OpcodeFamily};
+use super::info::{OpcodeInfo, opcode_info};
+
+pub(super) const OPCODES: &[OpcodeInfo] = &[
+    opcode_info(
+        Opcode::Nop,
+        OpcodeFamily::Core,
+        "nop",
+        OperandShape::None,
+        0x00,
+    ),
+    opcode_info(
+        Opcode::Trap,
+        OpcodeFamily::Core,
+        "trap",
+        OperandShape::String,
+        0x01,
+    ),
+    opcode_info(
+        Opcode::Pop,
+        OpcodeFamily::Core,
+        "pop",
+        OperandShape::None,
+        0x02,
+    ),
+    opcode_info(
+        Opcode::Dup,
+        OpcodeFamily::Core,
+        "dup",
+        OperandShape::None,
+        0x03,
+    ),
+    opcode_info(
+        Opcode::Swap,
+        OpcodeFamily::Core,
+        "swap",
+        OperandShape::None,
+        0x04,
+    ),
+    opcode_info(
+        Opcode::LdUnit,
+        OpcodeFamily::Core,
+        "ld.unit",
+        OperandShape::None,
+        0x05,
+    ),
+    opcode_info(
+        Opcode::LdTrue,
+        OpcodeFamily::Core,
+        "ld.true",
+        OperandShape::None,
+        0x06,
+    ),
+    opcode_info(
+        Opcode::LdFalse,
+        OpcodeFamily::Core,
+        "ld.false",
+        OperandShape::None,
+        0x07,
+    ),
+    opcode_info(
+        Opcode::LdNull,
+        OpcodeFamily::Core,
+        "ld.null",
+        OperandShape::Type,
+        0x08,
+    ),
+    opcode_info(
+        Opcode::LdC,
+        OpcodeFamily::Core,
+        "ld.c",
+        OperandShape::Constant,
+        0x09,
+    ),
+    opcode_info(
+        Opcode::LdCI4,
+        OpcodeFamily::Core,
+        "ld.c.i4",
+        OperandShape::I16,
+        0x0A,
+    ),
+    opcode_info(
+        Opcode::LdCI8,
+        OpcodeFamily::Core,
+        "ld.c.i8",
+        OperandShape::I16,
+        0x0B,
+    ),
+    opcode_info(
+        Opcode::LdCF4,
+        OpcodeFamily::Core,
+        "ld.c.f4",
+        OperandShape::I16,
+        0x0C,
+    ),
+    opcode_info(
+        Opcode::LdCF8,
+        OpcodeFamily::Core,
+        "ld.c.f8",
+        OperandShape::I16,
+        0x0D,
+    ),
+    opcode_info(
+        Opcode::LdStr,
+        OpcodeFamily::Core,
+        "ld.str",
+        OperandShape::String,
+        0x0E,
+    ),
+];
