+++ crates/music_seam/src/opcode/table/object.rs
@@ -1,0 +1,84 @@
+use crate::instruction::OperandShape;
+
+use super::super::{Opcode, OpcodeFamily};
+use super::info::{OpcodeInfo, opcode_info};
+
+pub(super) const OPCODES: &[OpcodeInfo] = &[
+    opcode_info(
+        Opcode::NewObj,
+        OpcodeFamily::Object,
+        "new.obj",
+        OperandShape::TypeLen,
+        0x70,
+    ),
+    opcode_info(
+        Opcode::NewArr,
+        OpcodeFamily::Object,
+        "new.arr",
+        OperandShape::TypeLen,
+        0x71,
+    ),
+    opcode_info(
+        Opcode::ArrNew,
+        OpcodeFamily::Object,
+        "arr.new",
+        OperandShape::Type,
+        0x72,
+    ),
+    opcode_info(
+        Opcode::LdElem,
+        OpcodeFamily::Object,
+        "ld.elem",
+        OperandShape::None,
+        0x73,
+    ),
+    opcode_info(
+        Opcode::StElem,
+        OpcodeFamily::Object,
+        "st.elem",
+        OperandShape::None,
+        0x74,
+    ),
+    opcode_info(
+        Opcode::LdElemA,
+        OpcodeFamily::Object,
+        "ld.elem.a",
+        OperandShape::Type,
+        0x75,
+    ),
+    opcode_info(
+        Opcode::LdLen,
+        OpcodeFamily::Object,
+        "ld.len",
+        OperandShape::None,
+        0x76,
+    ),
+    opcode_info(
+        Opcode::Slice,
+        OpcodeFamily::Object,
+        "slice",
+        OperandShape::None,
+        0x77,
+    ),
+    opcode_info(
+        Opcode::CpObj,
+        OpcodeFamily::Object,
+        "cp.obj",
+        OperandShape::Type,
+        0x78,
+    ),
+    opcode_info(
+        Opcode::DupObj,
+        OpcodeFamily::Object,
+        "dup.obj",
+        OperandShape::Type,
+        0x79,
+    ),
+    opcode_info(
+        Opcode::DropObj,
+        OpcodeFamily::Object,
+        "drop.obj",
+        OperandShape::Type,
+        0x7A,
+    ),
+];
