+++ crates/music_seam/src/opcode/table/scalar.rs
@@ -1,0 +1,238 @@
+use crate::instruction::OperandShape;
+
+use super::super::{Opcode, OpcodeFamily};
+use super::info::{OpcodeInfo, opcode_info};
+
+pub(super) const OPCODES: &[OpcodeInfo] = &[
+    opcode_info(
+        Opcode::Neg,
+        OpcodeFamily::Scalar,
+        "neg",
+        OperandShape::None,
+        0x20,
+    ),
+    opcode_info(
+        Opcode::Add,
+        OpcodeFamily::Scalar,
+        "add",
+        OperandShape::None,
+        0x21,
+    ),
+    opcode_info(
+        Opcode::Sub,
+        OpcodeFamily::Scalar,
+        "sub",
+        OperandShape::None,
+        0x22,
+    ),
+    opcode_info(
+        Opcode::Mul,
+        OpcodeFamily::Scalar,
+        "mul",
+        OperandShape::None,
+        0x23,
+    ),
+    opcode_info(
+        Opcode::AddOvf,
+        OpcodeFamily::Scalar,
+        "add.ovf",
+        OperandShape::None,
+        0x24,
+    ),
+    opcode_info(
+        Opcode::SubOvf,
+        OpcodeFamily::Scalar,
+        "sub.ovf",
+        OperandShape::None,
+        0x25,
+    ),
+    opcode_info(
+        Opcode::MulOvf,
+        OpcodeFamily::Scalar,
+        "mul.ovf",
+        OperandShape::None,
+        0x26,
+    ),
+    opcode_info(
+        Opcode::DivS,
+        OpcodeFamily::Scalar,
+        "div.s",
+        OperandShape::None,
+        0x27,
+    ),
+    opcode_info(
+        Opcode::DivU,
+        OpcodeFamily::Scalar,
+        "div.u",
+        OperandShape::None,
+        0x28,
+    ),
+    opcode_info(
+        Opcode::RemS,
+        OpcodeFamily::Scalar,
+        "rem.s",
+        OperandShape::None,
+        0x29,
+    ),
+    opcode_info(
+        Opcode::RemU,
+        OpcodeFamily::Scalar,
+        "rem.u",
+        OperandShape::None,
+        0x2A,
+    ),
+    opcode_info(
+        Opcode::And,
+        OpcodeFamily::Scalar,
+        "and",
+        OperandShape::None,
+        0x2B,
+    ),
+    opcode_info(
+        Opcode::Or,
+        OpcodeFamily::Scalar,
+        "or",
+        OperandShape::None,
+        0x2C,
+    ),
+    opcode_info(
+        Opcode::Xor,
+        OpcodeFamily::Scalar,
+        "xor",
+        OperandShape::None,
+        0x2D,
+    ),
+    opcode_info(
+        Opcode::Not,
+        OpcodeFamily::Scalar,
+        "not",
+        OperandShape::None,
+        0x2E,
+    ),
+    opcode_info(
+        Opcode::Shl,
+        OpcodeFamily::Scalar,
+        "shl",
+        OperandShape::None,
+        0x2F,
+    ),
+    opcode_info(
+        Opcode::ShrS,
+        OpcodeFamily::Scalar,
+        "shr.s",
+        OperandShape::None,
+        0x30,
+    ),
+    opcode_info(
+        Opcode::ShrU,
+        OpcodeFamily::Scalar,
+        "shr.u",
+        OperandShape::None,
+        0x31,
+    ),
+    opcode_info(
+        Opcode::Rotl,
+        OpcodeFamily::Scalar,
+        "rotl",
+        OperandShape::None,
+        0x32,
+    ),
+    opcode_info(
+        Opcode::Rotr,
+        OpcodeFamily::Scalar,
+        "rotr",
+        OperandShape::None,
+        0x33,
+    ),
+    opcode_info(
+        Opcode::Clz,
+        OpcodeFamily::Scalar,
+        "clz",
+        OperandShape::None,
+        0x34,
+    ),
+    opcode_info(
+        Opcode::Ctz,
+        OpcodeFamily::Scalar,
+        "ctz",
+        OperandShape::None,
+        0x35,
+    ),
+    opcode_info(
+        Opcode::Popcnt,
+        OpcodeFamily::Scalar,
+        "popcnt",
+        OperandShape::None,
+        0x36,
+    ),
+    opcode_info(
+        Opcode::Ceq,
+        OpcodeFamily::Scalar,
+        "ceq",
+        OperandShape::None,
+        0x37,
+    ),
+    opcode_info(
+        Opcode::Cne,
+        OpcodeFamily::Scalar,
+        "cne",
+        OperandShape::None,
+        0x38,
+    ),
+    opcode_info(
+        Opcode::CltS,
+        OpcodeFamily::Scalar,
+        "clt.s",
+        OperandShape::None,
+        0x39,
+    ),
+    opcode_info(
+        Opcode::CltU,
+        OpcodeFamily::Scalar,
+        "clt.u",
+        OperandShape::None,
+        0x3A,
+    ),
+    opcode_info(
+        Opcode::CgtS,
+        OpcodeFamily::Scalar,
+        "cgt.s",
+        OperandShape::None,
+        0x3B,
+    ),
+    opcode_info(
+        Opcode::CgtU,
+        OpcodeFamily::Scalar,
+        "cgt.u",
+        OperandShape::None,
+        0x3C,
+    ),
+    opcode_info(
+        Opcode::CleS,
+        OpcodeFamily::Scalar,
+        "cle.s",
+        OperandShape::None,
+        0x3D,
+    ),
+    opcode_info(
+        Opcode::CleU,
+        OpcodeFamily::Scalar,
+        "cle.u",
+        OperandShape::None,
+        0x3E,
+    ),
+    opcode_info(
+        Opcode::CgeS,
+        OpcodeFamily::Scalar,
+        "cge.s",
+        OperandShape::None,
+        0x40,
+    ),
+    opcode_info(
+        Opcode::CgeU,
+        OpcodeFamily::Scalar,
+        "cge.u",
+        OperandShape::None,
+        0x41,
+    ),
+];
