+++ crates/music_seam/src/opcode/table/types.rs
@@ -1,0 +1,112 @@
+use crate::instruction::OperandShape;
+
+use super::super::{Opcode, OpcodeFamily};
+use super::info::{OpcodeInfo, opcode_info};
+
+pub(super) const OPCODES: &[OpcodeInfo] = &[
+    opcode_info(
+        Opcode::LdType,
+        OpcodeFamily::Type,
+        "ld.type",
+        OperandShape::Type,
+        0x80,
+    ),
+    opcode_info(
+        Opcode::TypeOf,
+        OpcodeFamily::Type,
+        "type.of",
+        OperandShape::None,
+        0x81,
+    ),
+    opcode_info(
+        Opcode::IsInst,
+        OpcodeFamily::Type,
+        "is.inst",
+        OperandShape::Type,
+        0x82,
+    ),
+    opcode_info(
+        Opcode::Cast,
+        OpcodeFamily::Type,
+        "cast",
+        OperandShape::Type,
+        0x83,
+    ),
+    opcode_info(
+        Opcode::Conv,
+        OpcodeFamily::Type,
+        "conv",
+        OperandShape::Type,
+        0x84,
+    ),
+    opcode_info(
+        Opcode::ConvOvf,
+        OpcodeFamily::Type,
+        "conv.ovf",
+        OperandShape::Type,
+        0x85,
+    ),
+    opcode_info(
+        Opcode::ConvS,
+        OpcodeFamily::Type,
+        "conv.s",
+        OperandShape::Type,
+        0x86,
+    ),
+    opcode_info(
+        Opcode::ConvU,
+        OpcodeFamily::Type,
+        "conv.u",
+        OperandShape::Type,
+        0x87,
+    ),
+    opcode_info(
+        Opcode::ConvOvfS,
+        OpcodeFamily::Type,
+        "conv.ovf.s",
+        OperandShape::Type,
+        0x88,
+    ),
+    opcode_info(
+        Opcode::ConvOvfU,
+        OpcodeFamily::Type,
+        "conv.ovf.u",
+        OperandShape::Type,
+        0x89,
+    ),
+    opcode_info(
+        Opcode::Box,
+        OpcodeFamily::Type,
+        "box",
+        OperandShape::Type,
+        0x8A,
+    ),
+    opcode_info(
+        Opcode::Unbox,
+        OpcodeFamily::Type,
+        "unbox",
+        OperandShape::Type,
+        0x8B,
+    ),
+    opcode_info(
+        Opcode::SizeOf,
+        OpcodeFamily::Type,
+        "size.of",
+        OperandShape::Type,
+        0x8C,
+    ),
+    opcode_info(
+        Opcode::AlignOf,
+        OpcodeFamily::Type,
+        "align.of",
+        OperandShape::Type,
+        0x8D,
+    ),
+    opcode_info(
+        Opcode::LdTok,
+        OpcodeFamily::Type,
+        "ld.tok",
+        OperandShape::Type,
+        0x8E,
+    ),
+];
