+++ crates/music_seam/src/tests.rs
@@ -8,7 +8,9 @@ use crate::descriptor::{
 };
 use crate::instruction::{CodeEntry, Instruction, Label, Operand};
 use crate::opcode::Opcode;
-use crate::{AssemblyError, decode_binary, encode_binary, format_text, parse_text};
+use crate::{
+    AssemblyError, decode_binary, encode_binary, format_hil_projection, format_text, parse_text,
+};
 
 mod success {
     use super::*;
@@ -369,6 +371,23 @@ mod success {
                 ".native $main::puts param $Int result $Int abi \"c\" symbol \"puts\" cold"
             )
         );
+    }
+
+    #[test]
+    fn hil_projection_uses_profile_attribute_spelling() {
+        let mut artifact = Artifact::new();
+        let procedure_name = artifact.intern_string("main::work");
+        let _ = artifact.procedures.alloc(
+            ProcedureDescriptor::new(procedure_name, 0, 0, Box::new([]))
+                .with_hot(true)
+                .with_cold(true),
+        );
+
+        let projection = format_hil_projection(&artifact);
+        assert!(projection.contains("@profile(level := .hot)"));
+        assert!(projection.contains("@profile(level := .cold)"));
+        assert!(!projection.contains(concat!("@", "hot")));
+        assert!(!projection.contains(concat!("@", "cold")));
     }
 }
